+++ src/WriterModule/ev42/ev42_Writer.h
@@ -19,6 +19,7 @@ using FlatbufferMessage = FileWriter::FlatbufferMessage;
 
 class ev42_Writer : public WriterModule::Base {
 public:
+  ev42_Writer() : WriterModule::Base(true) {}
   void parse_config(std::string const &ConfigurationStream) override;
   InitResult init_hdf(hdf5::node::Group &HDFGroup,
                       std::string const &HDFAttributes) override;