+++ src/WriterModuleBase.h
@@ -30,7 +30,11 @@ enum class InitResult { ERROR = -1, OK = 0 };
 /// Example: Please see `src/schemas/ev42/ev42_rw.cpp`.
 class Base {
 public:
+  Base(bool AcceptRepeatedTimestamps)
+      : WriteRepeatedTimestamps(AcceptRepeatedTimestamps) {}
   virtual ~Base() = default;
+
+  bool acceptsRepeatedTimestamps() const { return WriteRepeatedTimestamps; }
 
   /// \brief Parses the configuration of a stream.
   ///
@@ -70,6 +74,9 @@ public:
   ///
   /// \param msg The message to process
   virtual void write(FileWriter::FlatbufferMessage const &Message) = 0;
+
+private:
+  bool WriteRepeatedTimestamps;
 };
 
 class WriterException : public std::runtime_error {